rename(common): buildEnsureAppReadyInput
